separate trusty storage api from implementation
