+++ trusty/trusty-service/trusty-service-common/pom.xml
@@ -46,6 +46,10 @@
       <artifactId>trusty-storage-api</artifactId>
     </dependency>
     <dependency>
+      <groupId>org.kie.kogito</groupId>
+      <artifactId>trusty-storage-common</artifactId>
+    </dependency>
+    <dependency>
       <groupId>io.quarkus</groupId>
       <artifactId>quarkus-smallrye-health</artifactId>
     </dependency>
