+++ trusty/trusty-storage/trusty-storage-common/pom.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xmlns="http://maven.apache.org/POM/4.0.0"
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <parent>
     <artifactId>trusty-storage</artifactId>
@@ -9,13 +9,17 @@
   </parent>
   <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>trusty-storage-api</artifactId>
-  <name>Kogito :: Trusty Storage API</name>
+  <name>Kogito :: Trusty Storage Common</name>
+  <artifactId>trusty-storage-common</artifactId>
 
   <dependencies>
     <dependency>
       <groupId>org.kie.kogito</groupId>
-      <artifactId>tracing-decision-api</artifactId>
+      <artifactId>trusty-storage-api</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.kie.kogito</groupId>
+      <artifactId>persistence-commons-api</artifactId>
     </dependency>
   </dependencies>
