+++ trusty/trusty-storage/trusty-storage-common/src/main/java/org/kie/kogito/trusty/storage/common/TrustyStorageServiceImpl.java
@@ -14,37 +14,35 @@
  * limitations under the License.
  */
 
-package org.kie.kogito.trusty.storage.api;
+package org.kie.kogito.trusty.storage.common;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.inject.Inject;
 
 import org.kie.kogito.persistence.api.Storage;
+import org.kie.kogito.persistence.api.StorageService;
 import org.kie.kogito.trusty.storage.api.model.DMNModelWithMetadata;
 import org.kie.kogito.trusty.storage.api.model.Decision;
 import org.kie.kogito.trusty.storage.api.model.ExplainabilityResult;
 
-public interface TrustyStorageService {
+@ApplicationScoped
+public class TrustyStorageServiceImpl implements TrustyStorageService {
 
-    String DECISIONS_STORAGE = "decisions";
-    String EXPLAINABILITY_RESULTS_STORAGE = "explainability-results";
-    String MODELS_STORAGE = "models";
+    @Inject
+    StorageService storageService;
 
-    /**
-     * Gets the decision storage.
-     *
-     * @return The Storage for decisions.
-     */
-    Storage<String, Decision> getDecisionsStorage();
+    @Override
+    public Storage<String, Decision> getDecisionsStorage() {
+        return storageService.getCache(DECISIONS_STORAGE, Decision.class);
+    }
 
-    /**
-     * Gets the explainability result storage.
-     *
-     * @return The Storage for explainability results.
-     */
-    Storage<String, ExplainabilityResult> getExplainabilityResultStorage();
+    @Override
+    public Storage<String, ExplainabilityResult> getExplainabilityResultStorage() {
+        return storageService.getCache(EXPLAINABILITY_RESULTS_STORAGE, ExplainabilityResult.class);
+    }
 
-    /**
-     * Gets the model definition storage.
-     *
-     * @return The Storage for model definitions.
-     */
-    Storage<String, DMNModelWithMetadata> getModelStorage();
+    @Override
+    public Storage<String, DMNModelWithMetadata> getModelStorage() {
+        return storageService.getCache(MODELS_STORAGE, DMNModelWithMetadata.class);
+    }
 }
