+++ trusty/trusty-storage/trusty-storage-infinispan/pom.xml
@@ -19,6 +19,10 @@
     </dependency>
     <dependency>
       <groupId>org.kie.kogito</groupId>
+      <artifactId>trusty-storage-common</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.kie.kogito</groupId>
       <artifactId>persistence-commons-infinispan</artifactId>
     </dependency>
