+++ trusty/trusty-storage/trusty-storage-redis/pom.xml
@@ -21,6 +21,10 @@
       <groupId>org.kie.kogito</groupId>
       <artifactId>trusty-storage-api</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.kie.kogito</groupId>
+      <artifactId>trusty-storage-common</artifactId>
+    </dependency>
 
     <dependency>
       <groupId>io.quarkus</groupId>
