+++ trusty/trusty-storage/trusty-storage-redis/src/test/java/org/kie/kogito/trusty/storage/redis/IndexProviderTest.java
@@ -21,9 +21,9 @@ import org.junit.jupiter.api.Test;
 import org.kie.kogito.persistence.redis.RedisClientManager;
 import org.mockito.Mockito;
 
-import static org.kie.kogito.trusty.storage.api.TrustyStorageServiceImpl.DECISIONS_STORAGE;
-import static org.kie.kogito.trusty.storage.api.TrustyStorageServiceImpl.EXPLAINABILITY_RESULTS_STORAGE;
-import static org.kie.kogito.trusty.storage.api.TrustyStorageServiceImpl.MODELS_STORAGE;
+import static org.kie.kogito.trusty.storage.common.TrustyStorageService.DECISIONS_STORAGE;
+import static org.kie.kogito.trusty.storage.common.TrustyStorageService.EXPLAINABILITY_RESULTS_STORAGE;
+import static org.kie.kogito.trusty.storage.common.TrustyStorageService.MODELS_STORAGE;
 
 public class IndexProviderTest {
